+++ src/compiler/symtab.c
@@ -348,7 +348,6 @@ void symtab_init(uint32_t capacity)
 	kw_at_return = KW_DEF("@return");
 	attribute_list[ATTRIBUTE_ALIGN] = kw_at_align;
 	attribute_list[ATTRIBUTE_ALLOW_DEPRECATED] = KW_DEF("@allow_deprecated");
-	attribute_list[ATTRIBUTE_ALWAYSINIT] = KW_DEF("@alwaysinit");
 	attribute_list[ATTRIBUTE_BENCHMARK] = KW_DEF("@benchmark");
 	attribute_list[ATTRIBUTE_BIGENDIAN] = KW_DEF("@bigendian");
 	attribute_list[ATTRIBUTE_BUILTIN] = KW_DEF("@builtin");
@@ -370,6 +369,7 @@ void symtab_init(uint32_t capacity)
 	attribute_list[ATTRIBUTE_LITTLEENDIAN] = KW_DEF("@littleendian");
 	attribute_list[ATTRIBUTE_LOCAL] = KW_DEF("@local");
 	attribute_list[ATTRIBUTE_MAYDISCARD] = KW_DEF("@maydiscard");
+	attribute_list[ATTRIBUTE_MUSTINIT] = KW_DEF("@mustinit");
 	attribute_list[ATTRIBUTE_NAKED] = KW_DEF("@naked");
 	attribute_list[ATTRIBUTE_NOALIAS] = KW_DEF("@noalias");
 	attribute_list[ATTRIBUTE_NODISCARD] = KW_DEF("@nodiscard");