+++ releasenotes.md
@@ -21,7 +21,7 @@
 - Removed deprecated `$assignable`.
 - Enums now no longer directly support `+` and `-` – use ordinals instead.
 - For enums, using `++` and `--` will step through enums with implicit wrap-around.
-- Add an `@alwaysinit` attribute to enforce zero-initialization of a type. #3094
+- Add a `@mustinit` attribute to enforce zero-initialization of a type. #3094
 
 ## 0.7.11 Change list